+++ tusas.cpp
@@ -149,8 +149,8 @@ int main(int argc, char *argv[])
     double dt = paramList.get<double> (TusasdtNameString);
 
     const int intnumSteps = paramList.get<int> (TusasntNameString);
-    int64_t numSteps = static_cast<int64_t>(intnumSteps);
-    const int64_t int64numSteps = paramList.get<int64_t> (Tusasnt64NameString);
+    long long numSteps = static_cast<long long>(intnumSteps);
+    const long long int64numSteps = paramList.get<long long> (Tusasnt64NameString);
     if( (int64numSteps > 0 ) && (0 ==intnumSteps)) numSteps = int64numSteps;
     
     timestep<double> * model;
@@ -172,7 +172,7 @@ int main(int argc, char *argv[])
     model->initialize();
     
     double curTime = model->get_start_time();
-    int64_t elapsedSteps = model->get_start_step();
+    long long elapsedSteps = model->get_start_step();
     double endTime = static_cast<double>(numSteps)*dt;//curTime + ((double)numSteps-elapsedSteps)*dt;
     
     bool timeadapt = paramList.get<bool>(TusasadaptiveTimestepNameString);
@@ -191,7 +191,7 @@ int main(int argc, char *argv[])
 	    << "  ( "<<(float)curTime/(float)endTime*100. <<" % )   t = "
 	    <<curTime<<"  t final = "<<endTime<< endl<<endl<<endl;
       }
-      int64_t outputfreq = static_cast<int64_t>(paramList.get<int> (TusasoutputfreqNameString));
+      long long outputfreq = static_cast<long long>(paramList.get<int> (TusasoutputfreqNameString));
       if(0 == outputfreq ) outputfreq = numSteps;
       if(0 == elapsedSteps%outputfreq &&
 	 curTime <= endTime &&
